pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.yapf]
based_on_style = "pep8"
indent_width = 4
column_limit = 100

[tool.isort]
line_length = 100

[project]
name = "resonate"
version = "1.0.0"
authors = [{ name = "Xiquan Li", email = "mtxiaoxi55@sjtu.edu.cn" }]
description = "Resonate is strong text-to-audio generator trained with online reinforcement learning"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  'torch >= 2.5.1',
  'huggingface_hub >= 0.26',
  'cython',
  'gitpython >= 3.1',
  'tensorboard >= 2.11',
  'numpy >= 1.21, <2.1',
  'Pillow >= 9.5',
  'opencv-python >= 4.8',
  'scipy >= 1.7',
  'tqdm >= 4.66.1',
  'gradio >= 3.34',
  'einops >= 0.6',
  'hydra-core >= 1.3.2',
  'requests',
  'torchdiffeq >= 0.2.5',
  'librosa >= 0.8.1',
  'nitrous-ema',
  'hydra_colorlog',
  'tensordict >= 0.6.1',
  'colorlog',
  'open_clip_torch >= 2.29.0',
  'av >= 14.0.1',
  'timm >= 1.0.12',
  'python-dotenv',
  'transformers == 4.57.3',
  'debugpy',
  'peft',
  'diffusers', 
  'accelerate', 
  'qwen_omni_utils',
  'audiobox_aesthetics', 
  'msclap',
  'matplotlib',
  'wandb'
]

[tool.hatch.build.targets.wheel]
packages = ["resonate"]
